Fix for WiMax 4G on at boot - SharedPreference setPersistant(true)
This should fix all of the 4G on at boot problems.  The earlier changes
are removed, and the setting is now saved in the SharedPreferences.
I guess they wanted 4G to always start at boot, because the
preference was using setPersistant(false).
I will submit a revert to the powerwidget, but it works as is.

Change-Id: Ic1abc3b04093425dcd6fcf2cca92955cc59df9cc