Better fix for 4G on at boot: CMWimaxSettings portion
Update: after more testing we don't need to test for the
setting, we only need to set it so other services know its
state.  This simply removes the unnecessary settings check.
It still works without this change, but it might be better
to do it to clean up the code.

Change-Id: I9c9122762ae04cb30a12fbacf0382850436c5153