Update patch set 9
Patch Set 9: Looks good to me, but someone else must approve

The earlier patch I submitted had the wimax_on setting in the CMWimax settings and WimaxButton. The frameworks/base portion moves the setting to where it should be in the WimaxHelper class, so the setting is only set if there is not an error in the service starting. This patch reverts the wimax_on setting in CMWimaxSettings. I have been testing without any issues.

Patch-set: 9