Update patch set 9
Patch Set 9: No score

Hi, I'm on supersonic here, the newest patch set needs the frameworks/base changes as well (Change Ibd0fd66d). In the patch that was merged I had placed the setting in the CMWimaxSettings and WimaxButton instead of the WimaxHelper.  The problem with the merged patch is that the setting should only be done after receiving the result that the service has successfully changed state.

Patch-set: 9